pyproject.toml
[project]
name = "web-novel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastapi[standard]>=0.120.0",
    "lxml>=6.0.2",
    "python-slugify[unidecode]>=8.0.4",
    "sqlalchemy>=2.0.44",
]
